Refactored inner classes out
