+++ service-search-nls/src/main/java/fi/nls/oskari/search/util/ScoredSearchHit.java
@@ -1,4 +1,16 @@
 package fi.nls.oskari.search.util;
 
-public class ScoredSearchHit {
+class ScoredSearchHit implements Comparable<ScoredSearchHit> {
+    public final int score;
+    public final String text;
+
+    public ScoredSearchHit(String t, int s) {
+        text = t;
+        score = s;
+    }
+
+    @Override
+    public int compareTo(ScoredSearchHit other) {
+        return score - other.score;
+    }
 }
